Selenium google map

+++ src/main/java/com/Repository/PlaceRepository.java
@@ -2,9 +2,7 @@ package com.Repository;
 
 import com.Model.Place;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Repository;
 
-public interface PlaceRepository extends CrudRepository<Place, Integer> {
+public interface PlaceRepository extends CrudRepository<Place, Long> {
 
 }
